Refactor no-op task runner and data loader to workspace
